+++ O365/address_book.py
@@ -4,9 +4,10 @@ from enum import Enum
 
 from dateutil.parser import parse
 
-from .message import Recipients, Message
+from .utils import Recipients
 from .utils import AttachableMixin, TrackerSet
 from .utils import Pagination, NEXT_LINK_KEYWORD, ApiComponent
+from .message import Message
 
 GAL_MAIN_RESOURCE = 'users'
