+++ O365/sharepoint.py
@@ -2,9 +2,9 @@ import logging
 
 from dateutil.parser import parse
 
+from .utils import ApiComponent
 from .address_book import Contact
 from .drive import Storage
-from .utils import ApiComponent
 
 log = logging.getLogger(__name__)
