+++ O365/utils/__init__.py
@@ -1,6 +1,7 @@
 from .attachment import BaseAttachments, BaseAttachment, AttachableMixin
 from .utils import ApiComponent, OutlookWellKnowFolderNames
 from .utils import ImportanceLevel, TrackerSet
+from .utils import Recipient, Recipients, HandleRecipientsMixin
 from .utils import NEXT_LINK_KEYWORD, ME_RESOURCE, USERS_RESOURCE
 from .utils import OneDriveWellKnowFolderNames, Pagination, Query
 from .token import BaseTokenBackend, Token, FileSystemTokenBackend, FirestoreBackend
